requirements.txt
aiohttp
arq
fastapi
google-cloud-storage
streaming-form-data
uvicorn